requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
numpy==1.26.3
plotly>=5.18.0
openpyxl>=3.1.2
seaborn==0.13.1
matplotlib==3.8.2
typing-extensions>=4.8.0
msal>=1.25.0
python-dotenv>=1.0.0
Office365-REST-Python-Client>=2.4.0
requests>=2.31.0
streamlit-oauth>=0.1.14